Revert previous commit and fix via obtaining Delegation Token to credentials
